+++ product/src/main/java/com/emotionalcart/product/presentation/ProductController.java
@@ -1,8 +1,8 @@
 package com.emotionalcart.product.presentation;
 
 import com.emotionalcart.product.application.ProductService;
+import com.emotionalcart.product.presentation.dto.ReadCategories;
 import com.emotionalcart.product.presentation.dto.ReadProductReviews;
-import com.emotionalcart.product.application.dto.ReadCategories;
 
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Page;
